+++ src/test/scala/cognite/spark/RawTableRelationTest.scala
@@ -1,8 +1,9 @@
 package cognite.spark
 
+import com.cognite.sdk.scala.v1.RawRow
 import com.fasterxml.jackson.databind.{DeserializationFeature, ObjectMapper}
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import io.circe.{Json, JsonObject}
+import io.circe.Json
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.types._
@@ -12,7 +13,6 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
   import RawTableRelation._
   import spark.implicits._
 
-  private val readApiKey = System.getenv("TEST_API_KEY_READ")
   private val writeApiKey = System.getenv("TEST_API_KEY_WRITE")
 
   private def collectToSet[A](df: DataFrame): Set[A] =
@@ -46,102 +46,113 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
     ("key6", 123, """{ "___key": "___k2", "value": 2, "__key": "__k2", "key": "k2" }""")
   )
 
-  private val dfWithoutLastChangedSchema = StructType(
-    Seq(StructField("notLastChanged", LongType, false), StructField("value", IntegerType, false)))
-  private val dfWithoutLastChangedData = Seq(
-    ("key1", 121, """{ "notLastChanged": 1, "value": 1 }"""),
-    ("key2", 122, """{ "notLastChanged": 2, "value": 2 }""")
+  private val dfWithoutlastUpdatedTimeSchema = StructType(
+    Seq(StructField("notlastUpdatedTime", LongType, false), StructField("value", IntegerType, false)))
+  private val dfWithoutlastUpdatedTimeData = Seq(
+    ("key1", 121, """{ "notlastUpdatedTime": 1, "value": 1 }"""),
+    ("key2", 122, """{ "notlastUpdatedTime": 2, "value": 2 }""")
   )
-  private val dfWithLastChangedSchema = StructType(
-    Seq(StructField("lastChanged", LongType, false), StructField("value", IntegerType, false)))
-  private val dfWithLastChangedData = Seq(
-    ("key3", 123, """{ "lastChanged": 1, "value": 1 }"""),
-    ("key4", 124, """{ "lastChanged": 2, "value": 2 }""")
+  private val dfWithlastUpdatedTimeSchema = StructType(
+    Seq(StructField("lastUpdatedTime", LongType, false), StructField("value", IntegerType, false)))
+  private val dfWithlastUpdatedTimeData = Seq(
+    ("key3", 123, """{ "lastUpdatedTime": 1, "value": 1 }"""),
+    ("key4", 124, """{ "lastUpdatedTime": 2, "value": 2 }""")
   )
-  private val dfWithManyLastChangedSchema = StructType(
+  private val dfWithManylastUpdatedTimeSchema = StructType(
     Seq(
-      StructField("lastChanged", LongType, true),
-      StructField("__lastChanged", LongType, false),
-      StructField("___lastChanged", LongType, false),
+      StructField("lastUpdatedTime", LongType, true),
+      StructField("__lastUpdatedTime", LongType, false),
+      StructField("___lastUpdatedTime", LongType, false),
       StructField("value", IntegerType, false)
     ))
-  private val dfWithManyLastChangedData = Seq(
-    ("key5", 125, """{ "___lastChanged": 111, "__lastChanged": 11, "value": 1 }"""),
-    (
-      "key6",
-      126,
-      """{ "___lastChanged": 222, "value": 2, "__lastChanged": 22, "lastChanged": 2 }""")
+  private val dfWithManylastUpdatedTimeData = Seq(
+    ("key5", 125, """{ "___lastUpdatedTime": 111, "__lastUpdatedTime": 11, "value": 1 }"""),
+    ("key6", 126, """{ "___lastUpdatedTime": 222, "value": 2, "__lastUpdatedTime": 22, "lastUpdatedTime": 2 }""")
   )
 
+  it should "smoke test raw" taggedAs WriteTest in {
+    val limit = 100
+    val partitions = 10
+    val df = spark.read.format("cognite.spark")
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("limit", limit)
+      .option("partitions", partitions)
+      .option("database", "testdb")
+      .option("table", "cryptoAssets")
+      .option("inferSchema", "true")
+      .option("inferSchemaLimit", "100")
+      .load()
+
+    df.createTempView("raw")
+    val res = spark.sqlContext
+      .sql("select * from raw")
+    assert(res.count == limit * partitions)
+  }
+
   "A RawTableRelation" should "allow data columns named key, _key etc. but rename them to _key, __key etc." in {
-    val dfWithoutKey = dfWithoutKeyData.toDF("key", "lastChanged", "columns")
+    val dfWithoutKey = dfWithoutKeyData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithoutKey =
       flattenAndRenameColumns(spark.sqlContext, dfWithoutKey, dfWithoutKeySchema)
     processedWithoutKey.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "notKey", "value"))
+      Set("key", "lastUpdatedTime", "notKey", "value"))
     collectToSet[String](processedWithoutKey.select($"key")) should equal(Set("key1", "key2"))
 
-    val dfWithKey = dfWithKeyData.toDF("key", "lastChanged", "columns")
+    val dfWithKey = dfWithKeyData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithKey = flattenAndRenameColumns(spark.sqlContext, dfWithKey, dfWithKeySchema)
-    processedWithKey.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "_key", "value"))
+    processedWithKey.schema.fieldNames.toSet should equal(Set("key", "lastUpdatedTime", "_key", "value"))
     collectToSet[String](processedWithKey.select($"key")) should equal(Set("key3", "key4"))
     collectToSet[String](processedWithKey.select($"_key")) should equal(Set("k1", "k2"))
 
-    val dfWithManyKeys = dfWithManyKeysData.toDF("key", "lastChanged", "columns")
+    val dfWithManyKeys = dfWithManyKeysData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithManyKeys =
       flattenAndRenameColumns(spark.sqlContext, dfWithManyKeys, dfWithManyKeysSchema)
     processedWithManyKeys.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "____key", "___key", "_key", "value"))
+      Set("key", "lastUpdatedTime", "____key", "___key", "_key", "value"))
 
     collectToSet[String](processedWithManyKeys.select($"key")) should equal(Set("key5", "key6"))
     collectToSet[String](processedWithManyKeys.select($"_key")) should equal(Set(null, "k2"))
     collectToSet[String](processedWithManyKeys.select($"___key")) should equal(Set("__k1", "__k2"))
-    collectToSet[String](processedWithManyKeys.select($"____key")) should equal(
-      Set("___k1", "___k2"))
+    collectToSet[String](processedWithManyKeys.select($"____key")) should equal(Set("___k1", "___k2"))
   }
 
-  it should "allow data columns named lastChanged, _lastChanged etc. but rename them to _lastChanged, __lastChanged etc." in {
-    val dfWithoutLastChanged = dfWithoutLastChangedData.toDF("key", "lastChanged", "columns")
-    val processedWithoutLastChanged =
-      flattenAndRenameColumns(spark.sqlContext, dfWithoutLastChanged, dfWithoutLastChangedSchema)
-    processedWithoutLastChanged.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "notLastChanged", "value"))
-    collectToSet[Long](processedWithoutLastChanged.select($"lastChanged")) should equal(
-      Set(121, 122))
+  it should "allow data columns named lastUpdatedTime, _lastUpdatedTime etc. but rename them to _lastUpdatedTime, __lastUpdatedTime etc." in {
+    val dfWithoutlastUpdatedTime = dfWithoutlastUpdatedTimeData.toDF("key", "lastUpdatedTime", "columns")
+    val processedWithoutlastUpdatedTime =
+      flattenAndRenameColumns(spark.sqlContext, dfWithoutlastUpdatedTime, dfWithoutlastUpdatedTimeSchema)
+    processedWithoutlastUpdatedTime.schema.fieldNames.toSet should equal(
+      Set("key", "lastUpdatedTime", "notlastUpdatedTime", "value"))
+    collectToSet[Long](processedWithoutlastUpdatedTime.select($"lastUpdatedTime")) should equal(Set(121, 122))
 
-    val dfWithLastChanged = dfWithLastChangedData.toDF("key", "lastChanged", "columns")
-    val processedWithLastChanged =
-      flattenAndRenameColumns(spark.sqlContext, dfWithLastChanged, dfWithLastChangedSchema)
-    processedWithLastChanged.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "_lastChanged", "value"))
-    collectToSet[Long](processedWithLastChanged.select($"lastChanged")) should equal(Set(123, 124))
-    collectToSet[Long](processedWithLastChanged.select($"_lastChanged")) should equal(Set(1, 2))
+    val dfWithlastUpdatedTime = dfWithlastUpdatedTimeData.toDF("key", "lastUpdatedTime", "columns")
+    val processedWithlastUpdatedTime =
+      flattenAndRenameColumns(spark.sqlContext, dfWithlastUpdatedTime, dfWithlastUpdatedTimeSchema)
+    processedWithlastUpdatedTime.schema.fieldNames.toSet should equal(
+      Set("key", "lastUpdatedTime", "_lastUpdatedTime", "value"))
+    collectToSet[Long](processedWithlastUpdatedTime.select($"lastUpdatedTime")) should equal(Set(123, 124))
+    collectToSet[Long](processedWithlastUpdatedTime.select($"_lastUpdatedTime")) should equal(Set(1, 2))
 
-    val dfWithManyLastChanged = dfWithManyLastChangedData.toDF("key", "lastChanged", "columns")
-    val processedWithManyLastChanged =
-      flattenAndRenameColumns(spark.sqlContext, dfWithManyLastChanged, dfWithManyLastChangedSchema)
-    processedWithManyLastChanged.schema.fieldNames.toSet should equal(
-      Set("key", "lastChanged", "____lastChanged", "___lastChanged", "_lastChanged", "value"))
+    val dfWithManylastUpdatedTime = dfWithManylastUpdatedTimeData.toDF("key", "lastUpdatedTime", "columns")
+    val processedWithManylastUpdatedTime =
+      flattenAndRenameColumns(spark.sqlContext, dfWithManylastUpdatedTime, dfWithManylastUpdatedTimeSchema)
+    processedWithManylastUpdatedTime.schema.fieldNames.toSet should equal(
+      Set("key", "lastUpdatedTime", "____lastUpdatedTime", "___lastUpdatedTime", "_lastUpdatedTime", "value"))
 
-    collectToSet[Long](processedWithManyLastChanged.select($"lastChanged")) should equal(
-      Set(125, 126))
-    collectToSet[Long](processedWithManyLastChanged.select($"_lastChanged")) should equal(
-      Set(null, 2))
-    collectToSet[Long](processedWithManyLastChanged.select($"___lastChanged")) should equal(
-      Set(11, 22))
-    collectToSet[Long](processedWithManyLastChanged.select($"____lastChanged")) should equal(
+    collectToSet[Long](processedWithManylastUpdatedTime.select($"lastUpdatedTime")) should equal(Set(125, 126))
+    collectToSet[Long](processedWithManylastUpdatedTime.select($"_lastUpdatedTime")) should equal(Set(null, 2))
+    collectToSet[Long](processedWithManylastUpdatedTime.select($"___lastUpdatedTime")) should equal(Set(11, 22))
+    collectToSet[Long](processedWithManylastUpdatedTime.select($"____lastUpdatedTime")) should equal(
       Set(111, 222))
   }
 
   it should "insert data with columns named _key, __key etc. as data columns key, _key, etc." in {
-    val dfWithKey = dfWithKeyData.toDF("key", "lastChanged", "columns")
+    val dfWithKey = dfWithKeyData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithKey = flattenAndRenameColumns(spark.sqlContext, dfWithKey, dfWithKeySchema)
     val (columnNames1, unRenamed1) = prepareForInsert(processedWithKey)
     columnNames1.toSet should equal(Set("key", "value"))
     collectToSet[String](unRenamed1.select("key")) should equal(Set("k1", "k2"))
 
-    val dfWithManyKeys = dfWithManyKeysData.toDF("key", "lastChanged", "columns")
+    val dfWithManyKeys = dfWithManyKeysData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithManyKeys =
       flattenAndRenameColumns(spark.sqlContext, dfWithManyKeys, dfWithManyKeysSchema)
     val (columnNames2, unRenamed2) = prepareForInsert(processedWithManyKeys)
@@ -151,22 +162,22 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
     collectToSet[String](unRenamed2.select("___key")) should equal(Set("___k1", "___k2"))
   }
 
-  it should "insert data with columns named _lastChanged, __lastChanged etc. as data columns lastChanged, _lastChanged, etc." in {
-    val dfWithLastChanged = dfWithLastChangedData.toDF("key", "lastChanged", "columns")
-    val processedWithLastChanged =
-      flattenAndRenameColumns(spark.sqlContext, dfWithLastChanged, dfWithLastChangedSchema)
-    val (columnNames1, unRenamed1) = prepareForInsert(processedWithLastChanged)
-    columnNames1.toSet should equal(Set("lastChanged", "value"))
-    collectToSet[Long](unRenamed1.select("lastChanged")) should equal(Set(1, 2))
+  it should "insert data with columns named _lastUpdatedTime, __lastUpdatedTime etc. as data columns lastUpdatedTime, _lastUpdatedTime, etc." in {
+    val dfWithlastUpdatedTime = dfWithlastUpdatedTimeData.toDF("key", "lastUpdatedTime", "columns")
+    val processedWithlastUpdatedTime =
+      flattenAndRenameColumns(spark.sqlContext, dfWithlastUpdatedTime, dfWithlastUpdatedTimeSchema)
+    val (columnNames1, unRenamed1) = prepareForInsert(processedWithlastUpdatedTime)
+    columnNames1.toSet should equal(Set("lastUpdatedTime", "value"))
+    collectToSet[Long](unRenamed1.select("lastUpdatedTime")) should equal(Set(1, 2))
 
-    val dfWithManyLastChanged = dfWithManyLastChangedData.toDF("key", "lastChanged", "columns")
-    val processedWithManyLastChanged =
-      flattenAndRenameColumns(spark.sqlContext, dfWithManyLastChanged, dfWithManyLastChangedSchema)
-    val (columnNames2, unRenamed2) = prepareForInsert(processedWithManyLastChanged)
-    columnNames2.toSet should equal(Set("lastChanged", "__lastChanged", "___lastChanged", "value"))
-    collectToSet[Long](unRenamed2.select($"lastChanged")) should equal(Set(null, 2))
-    collectToSet[Long](unRenamed2.select($"__lastChanged")) should equal(Set(11, 22))
-    collectToSet[Long](unRenamed2.select($"___lastChanged")) should equal(Set(111, 222))
+    val dfWithManylastUpdatedTime = dfWithManylastUpdatedTimeData.toDF("key", "lastUpdatedTime", "columns")
+    val processedWithManylastUpdatedTime =
+      flattenAndRenameColumns(spark.sqlContext, dfWithManylastUpdatedTime, dfWithManylastUpdatedTimeSchema)
+    val (columnNames2, unRenamed2) = prepareForInsert(processedWithManylastUpdatedTime)
+    columnNames2.toSet should equal(Set("lastUpdatedTime", "__lastUpdatedTime", "___lastUpdatedTime", "value"))
+    collectToSet[Long](unRenamed2.select($"lastUpdatedTime")) should equal(Set(null, 2))
+    collectToSet[Long](unRenamed2.select($"__lastUpdatedTime")) should equal(Set(11, 22))
+    collectToSet[Long](unRenamed2.select($"___lastUpdatedTime")) should equal(Set(111, 222))
   }
 
   val mapper: ObjectMapper = {
@@ -177,23 +188,23 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
     mapper
   }
 
-  "rowsToRawItems" should "return RawItems from Rows" in {
-    val dfWithKey = dfWithKeyData.toDF("key", "lastChanged", "columns")
+  "rowsToRawItems" should "return RawRows from Rows" in {
+    val dfWithKey = dfWithKeyData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithKey = flattenAndRenameColumns(spark.sqlContext, dfWithKey, dfWithKeySchema)
     val (columnNames, unRenamed) = prepareForInsert(processedWithKey)
-    val rawItems = rowsToRawItems(columnNames, unRenamed.collect.toSeq, mapper)
+    val rawItems: Seq[RawRow] = rowsToRawItems(columnNames, unRenamed.collect.toSeq, mapper)
     rawItems.map(_.key.toString).toSet should equal(Set("key3", "key4"))
 
-    val expectedResult: Seq[JsonObject] = Seq[JsonObject](
-      JsonObject(("key", Json.fromString("k1")), ("value", Json.fromInt(1))),
-      JsonObject(("key", Json.fromString("k2")), ("value", Json.fromInt(2)))
+    val expectedResult: Seq[Map[String, Json]] = Seq[Map[String, Json]](
+      Map(("key" -> Json.fromString("k1")), ("value" -> Json.fromInt(1))),
+      Map(("key" -> Json.fromString("k2")), ("value" -> Json.fromInt(2)))
     )
 
     rawItems.map(_.columns) should equal(expectedResult)
   }
 
   it should "throw an IllegalArgumentException when DataFrame has null key" in {
-    val dfWithKey = dfWithNullKeyData.toDF("key", "lastChanged", "columns")
+    val dfWithKey = dfWithNullKeyData.toDF("key", "lastUpdatedTime", "columns")
     val processedWithKey = flattenAndRenameColumns(spark.sqlContext, dfWithKey, dfWithKeySchema)
     val (columnNames, unRenamed) = prepareForInsert(processedWithKey)
     an[IllegalArgumentException] should be thrownBy rowsToRawItems(
@@ -207,6 +218,7 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
     val database = "testdb"
     val table = "future-event"
     val inferSchemaLimit = 1
+    val partitions = 10
     val df = spark.read
       .format("cognite.spark")
       .option("apiKey", writeApiKey)
@@ -214,6 +226,7 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
       .option("database", database)
       .option("table", table)
       .option("inferSchema", true)
+      .option("partitions", partitions)
       .option("metricsPrefix", metricsPrefix)
       .option("inferSchemaLimit", inferSchemaLimit)
       .option("collectSchemaInferenceMetrics", true)
@@ -225,10 +238,10 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
 
     val numRowsReadDuringSchemaInferance =
       getNumberOfRowsRead(metricsPrefix, s"raw.$database.$table.rows")
-    numRowsReadDuringSchemaInferance should be(inferSchemaLimit)
+    numRowsReadDuringSchemaInferance should be(inferSchemaLimit * partitions)
   }
 
-  "LastChanged" should "insert data without error" taggedAs (WriteTest) in {
+  "lastUpdatedTime" should "insert data without error" taggedAs (WriteTest) in {
     val destinationDf = spark.read
       .format("cognite.spark")
       .option("apiKey", writeApiKey)
@@ -243,14 +256,14 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
       .sql(s"""
           |select "key1" as key,
           |"foo" as columns,
-          |null as lastChanged
+          |null as lastUpdatedTime
       """.stripMargin)
       .select(destinationDf.columns.map(col): _*)
       .write
       .insertInto("destinationTable")
   }
 
-  it should "test that lastChanged filters are handled correctly" taggedAs (ReadTest) in {
+  it should "test that lastUpdatedTime filters are handled correctly" taggedAs (ReadTest) in {
     val df = spark.read
       .format("cognite.spark")
       .option("apiKey", writeApiKey)
@@ -258,8 +271,25 @@ class RawTableRelationTest extends FlatSpec with Matchers with SparkTest {
       .option("database", "testdb")
       .option("table", "future-event")
       .option("inferSchema", true)
+      .option("partitions", "5")
       .load()
-      .where(s"lastChanged > 1561117603000 and lastChanged < 1561117753000")
+      .where(s"lastUpdatedTime >= timestamp('2019-06-21 11:48:00.000Z') and lastUpdatedTime <= timestamp('2019-06-21 11:50:00.000Z')")
     assert(df.count() == 10)
+  }
+
+  it should "handle various numbers of partitions" taggedAs (ReadTest) in {
+    for(partitions <- Seq("1", "5", "10", "20")) {
+    val df = spark.read
+      .format("cognite.spark")
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("database", "testdb")
+      .option("table", "future-event")
+      .option("inferSchema", true)
+      .option("partitions", partitions)
+      .load()
+      .where(s"lastUpdatedTime >= timestamp('2019-06-21 11:48:00.000Z') and lastUpdatedTime <= timestamp('2019-06-21 11:50:00.000Z')")
+    assert(df.count() == 10)
+    }
   }
 }